api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
anthropic==0.40.0
google-cloud-speech==2.27.0
pydantic==2.9.0
websockets==13.0
pytest==8.3.0
pytest-asyncio==0.24.0
httpx==0.27.0
